+++ terraform/ces-module/main.tf
@@ -6,29 +6,33 @@ terraform {
       source  = "hashicorp/helm"
       version = ">=2.12.1"
     }
+    kubernetes = {
+      source  = "hashicorp/kubernetes"
+      version = "~> 2.30"
+    }
   }
 }
 
 locals {
   split_fqdn = split(".", var.ces_fqdn)
   # Top Level Domain extracted from fully qualified domain name. k3ces.local is used for development mode and empty fqdn.
-  tld        = var.ces_fqdn != "" ? "${element( split(".", var.ces_fqdn), length(local.split_fqdn) - 2)}.${element(local.split_fqdn, length(local.split_fqdn) - 1)}" : "k3ces.local"
-}
-
-
-provider "helm" {
-  kubernetes {
-    host                   = var.kubernetes_host
-    client_certificate     = var.kubernetes_client_certificate
-    client_key             = var.kubernetes_client_key
-    cluster_ca_certificate = var.kubernetes_cluster_ca_certificate
-  }
-
-  registry {
-    url      = "${var.helm_registry_schema}://${var.helm_registry_host}"
-    username = var.helm_registry_username
-    password = base64decode(var.helm_registry_password)
-  }
+  topLevelDomain = var.ces_fqdn != "" ? "${element( split(".", var.ces_fqdn), length(local.split_fqdn) - 2)}.${element(local.split_fqdn, length(local.split_fqdn) - 1)}" : "k3ces.local"
+  splitComponentNamespaces = [
+    for componentStr in var.components :
+    {
+      namespace = split("/", componentStr)[0]
+      rest      = split("/", componentStr)[1] //provoke error here, so that the build fails if no namespace or name is given
+    }
+  ]
+  parsedComponents = [
+    for namespaceAndRest in local.splitComponentNamespaces :
+    {
+      namespace = namespaceAndRest.namespace
+      name      = split(":", namespaceAndRest.rest)[0]
+      version   = length(split(":", namespaceAndRest.rest)) == 2 ? split(":", namespaceAndRest.rest)[1] : "latest"
+      deployNamespace = split(":", namespaceAndRest.rest)[0] != "k8s-longhorn" ? var.ces_namespace : "longhorn-system"
+    }
+  ]
 }
 
 resource "helm_release" "k8s-ces-setup" {
@@ -43,36 +47,37 @@ resource "helm_release" "k8s-ces-setup" {
   values = [
     templatefile("${path.module}/values.yaml.tftpl",
       {
-        "dogu_registry_endpoint"     = var.dogu_registry_endpoint
-        "dogu_registry_username"     = var.dogu_registry_username
-        "dogu_registry_password"     = var.dogu_registry_password
-        "dogu_registry_url_schema"   = var.dogu_registry_url_schema
-        "docker_registry_url"        = var.image_registry_url
-        "docker_registry_username"   = var.image_registry_username
-        "docker_registry_password"   = var.image_registry_password
-        "helm_registry_host"         = var.helm_registry_host
-        "helm_registry_schema"       = var.helm_registry_schema
-        "helm_registry_plain_http"   = var.helm_registry_plain_http
-        "helm_registry_insecure_tls" = var.helm_registry_insecure_tls
-        "helm_registry_username"     = var.helm_registry_username
-        "helm_registry_password"     = var.helm_registry_password
-        "setup_json"                 = yamlencode(templatefile(
+        "dogu_registry_endpoint"       = var.dogu_registry_endpoint
+        "dogu_registry_username"       = var.dogu_registry_username
+        "dogu_registry_password"       = var.dogu_registry_password
+        "dogu_registry_url_schema"     = var.dogu_registry_url_schema
+        "container_registry_secrets"   = var.container_registry_secrets
+        "helm_registry_host"           = var.helm_registry_host
+        "helm_registry_schema"         = var.helm_registry_schema
+        "helm_registry_plain_http"     = var.helm_registry_plain_http
+        "helm_registry_insecure_tls"   = var.helm_registry_insecure_tls
+        "helm_registry_username"       = var.helm_registry_username
+        "helm_registry_password"       = var.helm_registry_password
+        "component_operator_chart"     = var.component_operator_chart
+        "component_operator_crd_chart" = var.component_operator_crd_chart
+        "components"                   = local.parsedComponents
+        "setup_json"                   = yamlencode(templatefile(
           "${path.module}/setup.json.tftpl",
           {
             # https://docs.cloudogu.com/en/docs/system-components/ces-setup/operations/setup-json/
-            "admin_username"   = var.ces_admin_username,
-            "admin_password"   = var.ces_admin_password,
-            "admin_email"      = var.ces_admin_email,
-            "default_dogu"     = var.default_dogu,
-            "additional_dogus" = var.additional_dogus,
-            "fqdn" : var.ces_fqdn,
-            "domain" : local.tld
-            "certificateType" : var.ces_certificate_path == null ? "selfsigned" : "external"
-            "certificate" : var.ces_certificate_path != null ? replace(file(var.ces_certificate_path), "\n", "\\n") : ""
-            "certificateKey" : var.ces_certificate_key_path != null ? replace(file(var.ces_certificate_key_path), "\n", "\\n") : ""
+            "admin_username"  = var.ces_admin_username
+            "admin_password"  = var.ces_admin_password
+            "admin_email"     = var.ces_admin_email
+            "default_dogu"    = var.default_dogu
+            "dogus"           = var.dogus
+            "fqdn"            = var.ces_fqdn
+            "domain"          = local.topLevelDomain
+            "certificateType" = var.ces_certificate_path == null ? "selfsigned" : "external"
+            "certificate"     = var.ces_certificate_path != null ? replace(file(var.ces_certificate_path), "\n", "\\n") : ""
+            "certificateKey"  = var.ces_certificate_key_path != null ? replace(file(var.ces_certificate_key_path), "\n", "\\n") : ""
           }
         ))
-        "resource_patches" = var.resource_patches_file != null ? file("${path.root}/${var.resource_patches_file}") : ""
+        "resource_patches" = var.resource_patches
       })
   ]
 }